+++ geopackage-map/src/main/java/mil/nga/geopackage/map/tiles/overlay/GeoPackageOverlay.java
@@ -97,6 +97,16 @@ public class GeoPackageOverlay extends BoundedOverlay {
     }
 
     /**
+     * Get the tile retriever
+     *
+     * @return retriever
+     * @since 6.7.4
+     */
+    public TileRetriever getRetriever() {
+        return retriever;
+    }
+
+    /**
      * {@inheritDoc}
      */
     @Override